+++ Tests/UtiliKitTests.swift
@@ -6,6 +6,7 @@
 //
 
 import XCTest
+import CoreLocation
 @testable import UtiliKit
 
 class OpenSourceUtilitiesTests: XCTestCase {
@@ -76,5 +77,47 @@ class OpenSourceUtilitiesTests: XCTestCase {
         superview.layoutIfNeeded()
         
         XCTAssertEqual(superview.bounds.inset(by: insets), view.frame)
+    }
+
+    //MARK: - CLLocationCoordinate2D Tests
+    func test_CLLocationCoordinate2D_Valid() {
+        let coordinates = [CLLocationCoordinate2D(latitude: CLLocationDegrees(0),
+                                                  longitude: CLLocationDegrees(0)),
+                           CLLocationCoordinate2D(latitude: CLLocationDegrees(-0),
+                                                  longitude: CLLocationDegrees(-0)),
+                           CLLocationCoordinate2D(latitude: CLLocationDegrees(90),
+                                                  longitude: CLLocationDegrees(180)),
+                           CLLocationCoordinate2D(latitude: CLLocationDegrees(-90),
+                                                  longitude: CLLocationDegrees(180)),
+                           CLLocationCoordinate2D(latitude: CLLocationDegrees(90),
+                                                  longitude: CLLocationDegrees(-180)),
+                           CLLocationCoordinate2D(latitude: CLLocationDegrees(-90),
+                                                  longitude: CLLocationDegrees(-180))
+        ]
+
+        XCTAssertNil(coordinates.first { !$0.isValid })
+    }
+
+    func test_CLLocationCoordinate2d_Invalid() {
+        let coordinates = [CLLocationCoordinate2D(latitude: CLLocationDegrees(95),
+                                                  longitude: CLLocationDegrees(0)),
+                           CLLocationCoordinate2D(latitude: CLLocationDegrees(-95),
+                                                  longitude: CLLocationDegrees(0)),
+                           CLLocationCoordinate2D(latitude: CLLocationDegrees(0),
+                                                  longitude: CLLocationDegrees(185)),
+                           CLLocationCoordinate2D(latitude: CLLocationDegrees(0),
+                                                  longitude: CLLocationDegrees(-185)),
+                           CLLocationCoordinate2D(latitude: CLLocationDegrees(95),
+                                                  longitude: CLLocationDegrees(185)),
+                           CLLocationCoordinate2D(latitude: CLLocationDegrees(-95),
+                                                  longitude: CLLocationDegrees(185)),
+                           CLLocationCoordinate2D(latitude: CLLocationDegrees(95),
+                                                  longitude: CLLocationDegrees(-185)),
+                           CLLocationCoordinate2D(latitude: CLLocationDegrees(-95),
+                                                  longitude: CLLocationDegrees(-185))
+        ]
+
+        let invalidCoordinates = coordinates.filter { !$0.isValid}
+        XCTAssertTrue(coordinates.count == invalidCoordinates.count, "All test coordinates should be invalid")
     }
 }